CASTE-GUARD/backend/moderation-api/requirements.txt
fastapi
uvicorn
transformers
torch
pytesseract
opencv-python
Pillow
    